# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Video Streaming Service API
  version: 1.0.0
  contact:
    name: video-streaming-service
    url: https://github.com/Nuu-maan/video-streaming-service
  description: |
    Standalone JSON API for a video platform: accounts, uploads, transcoding,
    HLS/MP4 playback, comments, playlists, subscriptions, search, moderation
    and admin analytics. It is designed to be hosted on its own and consumed
    by a separate frontend (React, Next.js, Vue, ...) running on a different
    origin.

    ## The three things to know first

    1. **Auth is JWT Bearer.** Log in at `POST /auth/login` (the body field is
       `identifier` — a username *or* an email — plus `password`) and send
       `Authorization: Bearer <access_token>` on every authenticated call.
       Access tokens live 900 s; exchange the `refresh_token` at
       `POST /auth/refresh` for a new pair. A refresh token is *not* usable as
       an API credential, and an access token is rejected at `/auth/refresh`.

    2. **Every JSON response is wrapped in an envelope.**
       Success: `{"success": true, "data": ...}`.
       Lists (including `/search`): `{"success": true, "data": [...], "pagination": {"total", "page", "limit", "total_pages", "has_next", "has_previous"}}`.
       Errors: `{"success": false, "error": {"code": "...", "message": "..."}}`.
       Media routes (`.m3u8`, `.ts`, MP4, JPEG) return raw bytes, never the envelope.

...

    ## Other conventions

...
servers:
...
    variables:
...
security: []
paths:
...
  /videos/trending:
    get:
      tags: [Discovery]
      operationId: trending
      summary: Most engaged-with public videos inside a time window
      security: []
      parameters:
        - name: window
          in: query
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Plain array of videos (not paginated)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/VideoSearchItem"
        "400":
          description: Unknown `window` (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/videos/{id}/retry:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Admin]
      operationId: retryVideo
      summary: Re-queue a failed video for transcoding
      description: >-
        Requires `moderate_content`. Only a video in status `failed` may be
        retried; any other status is a 400.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Retry initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Video is not in status `failed` (`BAD_REQUEST`) or bad id (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/queue/stats:
    get:
      tags: [Admin]
      operationId: getQueueStats
      summary: Asynq default-queue statistics
      description: Requires `moderate_content`.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QueueStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/workers:
    get:
      tags: [Admin]
      operationId: listActiveWorkers
      summary: Active Asynq worker servers
      description: Requires `moderate_content`.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active workers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          workers:
                            type: array
                            items:
                              $ref: "#/components/schemas/WorkerInfo"
                          count:
                            type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <access_token>`. Obtain a token pair from
        `/auth/login` or `/auth/register`; renew with `/auth/refresh`.
        Access tokens live 900 s. A refresh token is not accepted here.
  parameters:
...
    VideoId:
      name: id
      in: path
      required: true
      description: Video id
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired or revoked access token (`UNAUTHORIZED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Authenticated, but lacking the required permission (`FORBIDDEN`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    NotFound:
      description: >-
        Resource not found (`NOT_FOUND`). Also the answer for a private video
        or playlist read by anyone but its owner — never 403.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Not found
  schemas:
...
    SuccessEnvelope:
      type: object
      description: 'Every non-media, non-list success response: `{"success": true, "data": ...}`'
      required: [success]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: The payload; its shape is documented per endpoint
...
    ErrorResponse:
      type: object
      description: 'Every error: `{"success": false, "error": {"code", "message"}}`'
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            Machine-readable error code. `AUTH_UNAVAILABLE` arrives with
            status 503 when the token-revocation store is unreachable and the
            server fails closed — retry with the same token rather than
            discarding credentials as you would on a 401.
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
            - ALREADY_EXISTS
            - USER_BANNED
            - DUPLICATE_REPORT
            - EMAIL_ALREADY_VERIFIED
            - INVALID_TOKEN
            - INVALID_CURRENT_PASSWORD
            - FILE_TOO_LARGE
            - INVALID_FORMAT
            - HLS_NOT_READY
            - PLAYLIST_NOT_FOUND
            - SEGMENT_NOT_FOUND
            - VIDEO_NOT_READY
            - FILE_NOT_FOUND
            - ALREADY_IN_PLAYLIST
            - AUTH_UNAVAILABLE
        message:
          type: string
          description: Human-readable explanation
    MessageResponse:
      description: A success envelope whose payload is a short confirmation
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
              additionalProperties: true

    # ── Enums ──
...
    VideoSearchItem:
      type: object
      properties:
        video_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        thumbnail_url:
          type: string
        duration:
          type: integer
          description: Seconds
        views:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        username:
          type: string
        user_id:
          type: string
          format: uuid
        user_avatar_url:
          type: string
        user_verified:
          type: boolean
        relevance:
          type: number
        snippet:
          type: string
...
    QueueStats:
      type: object
      properties:
        active:
          type: integer
        pending:
          type: integer
        scheduled:
          type: integer
        retry:
          type: integer
        archived:
          type: integer
        completed:
          type: integer
        aggregating:
          type: integer
        processed:
          type: integer
        failed:
          type: integer
        paused:
          type: boolean
        size:
          type: integer
    WorkerInfo:
      type: object
      properties:
        host:
          type: string
        pid:
          type: integer
        server_id:
          type: string
        concurrency:
          type: integer
        queues:
          type: object
          additionalProperties:
            type: integer
        started:
          type: string
          format: date-time
        active_tasks:
          type: integer
...